MythicBlockerFeature: Use `ContainerCloseEvent` instead of keybinds (#82)
This way, we don't need to check for any keybinds, as we are hooking into the container closing logic itself.
This commit also renames `EventFactory.onContainerClose` to `EventFactory.onClientboundContainerClosePacket` to highlight the difference between `ContainerCloseEvent` and `MenuCloseEvent`, as later is initiated by the server, and the new, `ContainerCloseEvent` is initiated by the client.