Merging with Config
Removed Overlay annotation
Made Config the main annotation and if it picks up an Overlay type as the field, it registers an OverlayHolder

Changed OverlayPosition to Overlay due to recent changes to the class and it storing everything you need for an overlay

Changed some comments for clearer, more descriptive comments.

Removed overlays package since it was merged with config package

Removed OverlayManage and Overlay annotation since they are now unnecessary.