fix: Patches Up SpeedBoostRedirector
Few flaws the original implementation had:

A) Didn't account for archers who had not fully unlocked Windy Feet, and could only do a 2 minutes boost (only accounted for 3 minutes boost).

B) This was also an uncolored pattern in disguise, as Wynncraft sends it with the same Aqua/Gray formatting you see in the getNotification method. This means that it actually just never worked at all.

I just made the regex work, as well as account for the 2/3 minutes distinction in both the pattern and the notification.