fix: Strict Coordinate Pattern now matches on five-digit long coords (#893)
This was an issue for both TNA and TCC, where both of those locations are really out there coordinate-wise and you needed to communicate locations (i.e. where is the portal in TCC, a dead player telling the rest of their party where the shadowling is in TNA), so let's loosen it up to prevent weird splitting.