refactor: Sort out position <-> location
We are now consistently using "location" to refer to integer coordinates (like on a map, or a BlockPos), and "position" to refer to real number coordinates (typically entity positions).

I have replaced as many instaces as possible by the Vanilla interface "Position". I have redesigned our "Location" so it only deals with integer coordinates.