feat: Remove @TypeOverride
Replace this by proper detection of generic types. This cannot be done in runtime on an object due to type erasure, but it can be extracted from the declaring field, which we actually check.

Also wrap the type so declared abstract collection types are replaced by concrete implementations suitable for json serialization.