feat: Enum improvements (#1390)
* refactor: Rename all enum to follow the UPPER_CAMEL name rule

* fix: The config UI reset button has stopped working

* * Change json for enum to use lowerCamel
* Introduce Upfixer to convert enum format
* Modify Upfixer framework to support use case
* Improve upfixer naming
* Use nice name of enums in settings UI

* Support enums nicely with /config command

* feat: More robust handling of null in config file

If the config file accidentally contains a null value, it will now be replaced by the default value, instead of causing a crash later on.

* Use a TypeAdapterFactory instead, to reach embedded enums

* feat: Remove @TypeOverride

Replace this by proper detection of generic types. This cannot be done in runtime on an object due to type erasure, but it can be extracted from the declaring field, which we actually check.

Also wrap the type so declared abstract collection types are replaced by concrete implementations suitable for json serialization.

* Fix merge error

* Remove allowNull annotation

* Qodana is right...

* Fix enum upfixer to work with embedded enums as well

* JsonElement.equals is stupid

* How did that end up committed? :-o

* Get rid of getClassOfConfigField()

* feat: Make reset button grayed-out if no change has been made (#1406)

This makes it easy to spot if a setting is left at default or has been changed by the user

* ci: spotless formatting

* Fix merge errors

---------

Co-authored-by: Kristof Kovacs <49001742+kristofbolyai@users.noreply.github.com>
Co-authored-by: magicus <magicus@users.noreply.github.com>