Added toggle option for territories on map
Added a config setting to choose whether the control key on the main map should act as a hold or a toggle for guild territories, as laid out in issue #1550. If the config setting is set to the default (true), the behavior should be the same as it is currently. If it is set to false, it will act as a toggle. There might be a better way to implement this than what I figured out, but this seems like it will work.