feat: Parse item prices lazily, only in the trade market
It should be pretty easy to make this work on other GUIs as well, if needed. Until that, price parsing can stay in the trade market model.
I am not fully happy with passing the itemStack to stat providers, however, it allows lazy-loading of prices (or anything else in the future). I am happy to take that performance over marginally better code structure.