feat: Parse item prices lazily, add AmountStatProvider and TotalPriceStatProvider (#2191)
* chore: Revert price parsing in GameItem

* feat: Parse item prices lazily, only in the trade market

It should be pretty easy to make this work on other GUIs as well, if needed. Until that, price parsing can stay in the trade market model.
I am not fully happy with passing the itemStack to stat providers, however, it allows lazy-loading of prices (or anything else in the future). I am happy to take that performance over marginally better code structure.

* feat: Add AmountStatProvider and TotalPriceStatProvider

* chore: review fixes

* chore: Rename to WynnItemData

* chore: Add itemstack to WynnItemData and use it

* chore: remove else

---------

Co-authored-by: Magnus Ihse Bursie <mag@icus.se>