feat: Make the content book much faster (#2203)
* feat: Add a container content verification to ContainerQueryHandler

This change allows the query handler to process set slot packets, instead of only processing set content packets. This is useful in the cases where wynn only sends set slot packets to update slots. Behavior is unchanged for any query that does not explicitly implement this new verification.

* fix: Fix WynntilsMenuScreen running container query in constructor

The query should be run in the first init, otherwise any references to the class will cause unnecessary scans...

* feat: Introduce backwards scrolling for filters in content book queries

This makes all content book queries a lot faster, sometimes saving even 10 steps (in common cases too, mini-quest -> quest takes 10 steps if you only go forwards)...

* chore: Only scan content book on the first init

No need to run queries on screen resizes.

* chore: Don't scan mini-quests every time

Most of the time the quest book is used for main quests, so this is a good tradeoff for speed.

* fix: Don't queue a query if it is already running

* fix: Catch all exceptions in ContainerQueryHandler

Even if we get a different exception than ContainerQueryException, we should still not get stuck in a bad state...

* feat: Add a config to cancel running queries on screen close

This is expected behavior for some users, but some prefer queries to finish in the background, especially tracking. The config should make it work both ways.