refactor: Make TooltipModel and all of its components generic
This refactor prepares the tooltip model to add custom tooltips to other types of items that are identifiable, like tomes. If a new type of identifiable, non-gear like item is added, like tomes and charms, the model can be adapted to support the new item type, simply by implementing an `IdentifiableItemInfo` for the type.