chore: As a prerequisite to parsing charm info, parse tome base stats correctly
Tome base stats are a special kind of identifications. Essentially they behave just like normal identifications, except that they always have a pre-identified, static value, and they are displayed in a non-regular manner. This kind of info is not used at the moment, but they will be used in the tome guide, and when sharing tomes in chat, with the new gear encoding.

Charms have this kind of stats as well, except that they display it like normal stats, and are variable.

See https://i.imgur.com/AD8O0y2.png for an example.