refactor: Treat SkillStatType as a regular stat
This refactor is not only needed for the gear encoding, I think it makes a lot of sense. I can see why skills were handled specially, as a fixed stat, but I think it is really obvious from this PR that it needed too much special handling, for no real benefit. Now that the V3 API makes it easy to tell the difference between "pre-identified" and not "pre-identified" stats, we can move skill stats back to every other stat. (Legacy API only had a difference between "pre-identified" items)