feat: Parse charms as gears, show charm identification percentages  (#2271)
* chore: As a prerequisite to parsing charm info, parse tome base stats correctly

Tome base stats are a special kind of identifications. Essentially they behave just like normal identifications, except that they always have a pre-identified, static value, and they are displayed in a non-regular manner. This kind of info is not used at the moment, but they will be used in the tome guide, and when sharing tomes in chat, with the new gear encoding.

Charms have this kind of stats as well, except that they display it like normal stats, and are variable.

See https://i.imgur.com/AD8O0y2.png for an example.

* chore: Rename CharmInfo to DeprecatedCharmInfo

* feat: Load charm info from API

* feat: Make it work with id parsing and ItemStatInfoFeature

* chore: remove DeprecatedCharmInfo

* fix: Fix bug in variable stat deserialization causing ranges to be calculated for items with fixed stats

* chore: remove unused method

* refactor: In preparation for charm identifications, add support for custom stat type ranges, rework the inverted stat system, other fixes to internal rolls (#2282)

* refactor: Allow custom base ranges for items, rework inverted calculation, add option to display stat as inverted

* fix: Fix SpellStatType internal names for percentage values

* fix: Fix elemental defense stats

Turns out the API was wrong, and the removed type does not exist, and is only a typo. This is now confirmed to work in-game.

* fix: Correctly calculate internal role in every case

I've done throughout testing, but I think just running with the assertions enabled without any issues proves that this should work.

* feat: Reworked perfect/increase/decrease percentages

* fix: fix legacy order not displaying skill stats (#2283)

* chore: [auto-generated] Update urls.json (#2284)

Co-authored-by: magicus <magicus@users.noreply.github.com>

* feat: Pet menu search (#2285)

* feat: Pet menu search

* ci: spotless formatting

---------

Co-authored-by: DonkeyBlaster <DonkeyBlaster@users.noreply.github.com>

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: magicus <magicus@users.noreply.github.com>
Co-authored-by: Ryan <57310593+DonkeyBlaster@users.noreply.github.com>
Co-authored-by: DonkeyBlaster <DonkeyBlaster@users.noreply.github.com>

* fix: Properly calculate charm stats

* fix: fix inconsistent elemental defence key handling

* refactor: rethink how treatAsNegative works

We should not modify the internal roll, as it causes issues in the future, for gear encoding. Also handle star calculations.

* chore: remove old note in CharmInfo

* fix: fix a bug with displayAsInverted/treatAsInverted in StatType

* chore: remove one more old comment

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: magicus <magicus@users.noreply.github.com>
Co-authored-by: Ryan <57310593+DonkeyBlaster@users.noreply.github.com>
Co-authored-by: DonkeyBlaster <DonkeyBlaster@users.noreply.github.com>