fix: Fix TestRegex failing tests
I am wondering if we should pass strings through StyledText, and test those against the patterns as well. We still have a lot of reset characters in these patterns, and 99% percent of them could be removed... However, I am not sure if it is worth the effort; if it works, don't fix it.