fix: Use proper hooks when enabling/disabling Discord
Previously we called `DiscordService#load` for every config change, even if it was necessary. This is partially because `DiscordService#isReady` will always be false if Discord is closed. These new hooks will run iff we changed the enabled/disabled state of the feature (+on startup, if the feature is enabled). This makes config related operations much faster (as I noticed that `/config` commands hang the game for seconds...).