fix: Use ticks for measuring slow packet difference, fix tick scheduler manager having an inconsistent order with the tick events
The logic was correct, but using `System.currentTimeMillis` is too "inaccurate" for tick timing in many cases. Since the packet difference is actually measured in ticks, we might as well use that here as well.
I've bumped the timeout ticks from 10 to 20, since the previous value was too low for some cases.

This also fixes the tick scheduler manager having an inconsistent order with the tick events, which also caused the same issue.
The new order of operations is:
1. TickAlwaysEvent
2. Run the tick scheduler (highest TickAlwaysEvent priority)
3. TickEvent

This fixes the issue where tick events could schedule a task for the next tick, but the tick scheduler would run it at the same tick event, since the tick always event was run after the tick event.