fix: Separate queued/immediate commands, execute queued commands when we are safely outside of dialogues
Previously we executed commands like /friend list in dialogues, which sent us a background message, so we couldn't parse it, and it appeared as spam to the user. I've separated "queued"/immediate commands and made sure queued commands are executed whenever we are outside a dialogue. Immediate commands, like chat commands, aliases, /class, /hub are still ran immediately (which should have been the case previously).