feat: Add support for events in string-parsed StyledTexts [skip ci] (#2365)
* feat: Add support for events in string-parsed StyledTexts [skip ci]

This should allow us to reconstruct fully working clickable/hoverable components after translation.

Also reworked how events are collected. Previously, it was only collected, when we ran `getString(INCLUDE_EVENTS)`.

I can't justify doing that, it was really stupid. In fact, I am glad I had to be the one to track it down. To be clear, this caused the events not be collected, unless a text was printed as text, which the `toString` method gladly does. This left me wondering why debug mode works, and production-run does not. Let me tell you the obvious: the debugger runs `toString`...

* chore: add a better test

* fix: Make events inherit like other style elements

Previously we've always explicitly added events as "style codes". However, this should not be the case, as we can use the same logic as we do with other style elements, like bold and italic, making events inherited (and reset by ChatFormatting.RESET or a new color).