fix: Fix item record crashing if there is a faulty item, fix item saving mode (#2386)
fix: Fix item record crashing if there is a faulty item, fix item save mode

This PR adds the proper save flags for saved items, which is extended encoding. We did not do this previously, slowly breaking the saved items with API changes. Unfortunately, we cannot make the item record keep up with item changes, all items in there need to be saved as-is at the time of recording. I've also added a fallback cleanup mechanism for the future, quarantining faulty items and trying to re-add items that parse (after a patch or other change).