feat: First take on derived map visibility
In theory, this works. We'll have to discuss if this is what we want. A design change I've applied for map visibility is not to return null values, but just return a default visibility. If we should return ALWAYS or the DEFAULT value is up to debate, but in-code DEFAULT makes more sense.

As for DerivedMapVisibility, we can likely instruct GSON to not serialize the null fields easily. If we can't, we'll add a custom deserializer.