refactor: Move ChatMessageReceivedEvent and ClientsideMessageEvent to StyledText (only) [skip ci] (#2684)
refactor: Move ChatMessageReceivedEvent and ClientsideMessageEvent to StyledText (only)

Previously, I did not do this move, as I wanted the event to provide a way to access the original component. However, the original component is not used anywhere, and the component itself is overridden with StyledText generated components in multiple places.