feat: Add better support for confirmationless dialogues
This PR should make non-standard NPC messages be parsed in confirmationless dialogues, which should provide a better overall dialogue support, with less flickering, and better translations.

This PR was tested with two quests, namely _Hollow Serenity_ and _Hero of Gavel_. Both of these quests have a wide variety of dialogues, and from my testing, there are no regressions. Hollow Serenity has some specific issues with selection dialogues with some NPCs, however they are not regressions of this PR, so no change was made to better support them.

Reviewing this PR is hard, as the dialogue system is complex with many edge-cases, so finding faults in logic is basically impossible. For this reason, I ask that reviewers try the new parsing, and let me know about their findings. Obviously, we can't test everything, so if we only look at a few quests, but we don't find any _new_ issues, then we may very well release this to the public.