fix: Try to parse a crafted items' name from the item, if it's not in the lore
There seems to be a one off edge-case, where the name is not set in a crafted item's lore, but is set as a custom name on the item itself.
This PR also fixes crafted items being parsed, when the item's name + effectStrength is not getting parsed, effectively leaving the parsed item in a buggy state.