wip: feat: Introduce graphed, dependency based downloads
`DownloadManager` has the documentation I'd put in the PR description.

TLDR; The goal is to eliminate download dependency race once and for all, all while taking benefit of having a manager behind all downloads, which include proper failure handling / retries, maximum parallel download settings, a clear and easily calculated download progress percentage and a list of download failures which can be shown to the user.