feat: Handle hash conflicts in url files
This PR aims to handle hash conflicts in the 3 versions of our URL list in a way that we always aim to re-download the data, if we are not "sure" our hash is correct. During the PR, I also discovered that loading data sources in constructors are VERY race-prone, as it's very likely the constructors run before the remote urls finished downloading, essentially limiting the mod to only use the local cache/bundled urls.json file.