feat: Redirect dropped item messages
Redirect the message that appears when an item is obtained via
harvesting a resource node or retrieving an item from a crafting station
when there is no space in inventory and the item is being dropped.
Contrary to what the pattern might make it seem, this message appears
per dropped item. If there are multiple items being dropped at the same
time (e. g. when exiting a crafting station with no inventory space and
16 crafted items) it will appear per item.

Signed-off-by: Christoph Matthias Kohnen <mail@cmkohnen.de>