fix: Make MapXAttributesImpl have custom deserializers instead of MapFeatures to avoid type issues (#3067)
* fix: Make MapXAttributesImpl have custom deserializers instead of MapFeatures to avoid type issues

This is the best of two worlds, now the MapFeatureImpl classes can be extended safely whenever we want a serializable/deserializable type, without having to worry that their type is always MapXImpl (as that is the best we could do with GSON's type hierarchy adapters). With this approach, we get to safely and correctly deserialize attributes.

This change does remove the old feature deserializer's featureId -> id and categoryId -> category mapping, which I actually find to be okay, as now the json representation is even closer to the classes themselves.

* fix: Add attribute serializer