fix(config): get UUID later for config filename
Some auth mods (namely Lunar Client's auth manager) will start the game
with a blank session and update it later in the launch sequence, late
enough that ConfigManager has already been loaded and userConfigFile
set. This changes the getter to lazily set the value to the current
session, making the config save on Lunar.

This might not be the best approach but it fixes the big issue of the
settings not saving between launches. Better handling of session changes
may be required.