feat: Handle the new HUD parts in Wynncraft 2.1.2_5
With this PR, we should correctly handle all the new (and old) HUD segments. The XP bar is no longer updated, rather, it is a new type of HUD. It is also less accurate now, as it now only has 53 possible states, whereas the old bar can have any float value. I've done basic adaptations for the profession experience segments, but they are as inaccurate as combat experience, so it is wise to stay with our current methods of character info scanning and parsing on harvest.