ci: Autopromote 'release' tagged PRs and allow promoting by hand (#3283)
Now, we don't need to add [skip ci] on every PR, rather, we can add a tag to the PR that we want to make a release with. This way, we don't need to filter out [release] from the PR title, and tags are much more visible and filterable. I also separated beta and release scripts to allow us to run them by hand separately.